Moved getThemeDataFromPlugins
#1660
