+++ packages/plugin/package.json
@@ -27,6 +27,7 @@
     "@deephaven/icons": "file:../icons",
     "@deephaven/iris-grid": "file:../iris-grid",
     "@deephaven/jsapi-types": "file:../jsapi-types",
+    "@deephaven/log": "file:../log",
     "@deephaven/react-hooks": "file:../react-hooks",
     "@fortawesome/fontawesome-common-types": "^6.1.1",
     "@fortawesome/react-fontawesome": "^0.2.0"
