+++ packages/plugin/src/PluginUtils.tsx
@@ -1,7 +1,17 @@
 import { isValidElement } from 'react';
-import { vsPreview } from '@deephaven/icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { type PluginModule, isWidgetPlugin } from './PluginTypes';
+import { getThemeKey, type ThemeData } from '@deephaven/components';
+import { vsPreview } from '@deephaven/icons';
+import Log from '@deephaven/log';
+import {
+  type PluginModule,
+  isWidgetPlugin,
+  PluginModuleMap,
+  ThemePlugin,
+  isThemePlugin,
+} from './PluginTypes';
+
+const log = Log.module('@deephaven/plugin.PluginUtils');
 
 export function pluginSupportsType(
   plugin: PluginModule | undefined,
@@ -31,4 +41,38 @@ export function getIconForPlugin(plugin: PluginModule): React.ReactElement {
   }
 
   return <FontAwesomeIcon icon={icon} />;
+}
+
+/**
+ * Extract theme data from theme plugins in the given plugin map.
+ * @param pluginMap
+ */
+export function getThemeDataFromPlugins(
+  pluginMap: PluginModuleMap
+): ThemeData[] {
+  const themePluginEntries = [...pluginMap.entries()].filter(
+    (entry): entry is [string, ThemePlugin] => isThemePlugin(entry[1])
+  );
+
+  log.debug('Getting theme data from plugins', themePluginEntries);
+
+  return themePluginEntries
+    .map(([pluginName, plugin]) => {
+      // Normalize to an array since config can be an array of configs or a
+      // single config
+      const configs = Array.isArray(plugin.themes)
+        ? plugin.themes
+        : [plugin.themes];
+
+      return configs.map(
+        ({ name, baseTheme, styleContent }) =>
+          ({
+            baseThemeKey: `default-${baseTheme ?? 'dark'}`,
+            themeKey: getThemeKey(pluginName, name),
+            name,
+            styleContent,
+          }) as const
+      );
+    })
+    .flat();
 }
